requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
icrawler==0.6.6
idna==3.4
lxml==4.9.2
magic-filter==1.0.9
multidict==6.0.4
Pillow==9.5.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
structlog==23.1.0
urllib3==1.26.15
uvloop==0.17.0
yarl==1.8.2
